+++ lib/src/TransactionUtil.java
@@ -315,7 +315,6 @@ public class TransactionUtil
       }
     }
 
-
     return tx.build();
     
   }